Consistent Kibana service start detection
FileSystemWatcher is flaky in detecting changes flushed to the Kibana log file; it is not notified of changes, but if the directory in which the log file exists is opened in explorer, the FileSystemWatcher detects the file changes. Replaced this with a polling interval that checks the size of the file since last checking, and reads any lines that have been written. This appears to be more robust,

Better progress messages for Kibana installation

Remove Kibana plugins as part of uninstallation

Derive PluginStateProviders for the separate products